requirements.txt
python-telegram-bot==20.7
python-dotenv>=1.0.0
aiosqlite>=0.19.0
httpx>=0.25.0
python-dateutil>=2.8.2
APScheduler>=3.10.0
uvicorn>=0.23.0
requests>=2.31.0
python-telegram-bot[job-queue]
Flask==2.3.2
gunicorn
